requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
nltk==3.8.1
vaderSentiment==3.3.2
google_play_scraper
psycopg2-binary==2.9.10
python-dotenv==1.0.0
matplotlib==3.8.0
seaborn==0.12.2
flake8==6.0.0
pytest==7.4.0
